requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyparsing==3.0.7
pytz==2022.1
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
